requirements.txt
fastapi
uvicorn
streamlit
requests
groq
python-dotenv
python-multipart
openai-whisper
pydantic
streamlit-mic-recorder

